+++ TDEngine2/include/graphics/CStaticMeshContainer.h
@@ -8,6 +8,7 @@
 
 
 #include "IStaticMeshContainer.h"
+#include "IMesh.h"
 #include "../ecs/CBaseComponent.h"
 
 
@@ -82,6 +83,8 @@ namespace TDEngine2
 
 			TDE2_API void SetSubMeshId(const std::string& meshName) override;
 
+			TDE2_API void SetSubMeshRenderInfo(const TSubMeshRenderInfo& info) override;
+
 			/*!
 				\brief The method sets up an internal handle which points to pair
 				vertex and index buffers within system that renders static meshes
@@ -91,6 +94,12 @@ namespace TDEngine2
 			*/
 
 			TDE2_API void SetSystemBuffersHandle(U32 handle) override;
+
+			TDE2_API void SetDirty(bool value) override;
+
+#if TDE2_EDITORS_ENABLED
+			TDE2_API void AddSubmeshIdentifier(const std::string& submeshId) override;
+#endif
 
 			/*!
 				\brief The method returns an identifier of used material
@@ -110,20 +119,36 @@ namespace TDEngine2
 
 			TDE2_API const std::string& GetSubMeshId() const override;
 
+			TDE2_API const TSubMeshRenderInfo& GetSubMeshInfo() const override;
+
 			/*!
 				\brief The method returns an internal handle which points to pair
 				vertex and index buffers within system that renders static meshes
 			*/
 
 			TDE2_API U32 GetSystemBuffersHandle() const override;
+
+			TDE2_API bool IsDirty() const override;
+
+#if TDE2_EDITORS_ENABLED
+			TDE2_API const std::vector<std::string>& GetSubmeshesIdentifiers() const override;
+#endif
 		protected:
 			DECLARE_INTERFACE_IMPL_PROTECTED_MEMBERS(CStaticMeshContainer)
 		protected:
-			std::string mMaterialName;
-			std::string mMeshName; /// \todo replace with GUID or something like that
-			std::string mSubMeshId = Wrench::StringUtils::GetEmptyStr(); ///< If the field's value is empty the whole mesh will be rendered with same material
+			std::string              mMaterialName;
+			std::string              mMeshName; /// \todo replace with GUID or something like that
+			std::string              mSubMeshId = Wrench::StringUtils::GetEmptyStr(); ///< If the field's value is empty the whole mesh will be rendered with same material
 
-			U32         mSystemBuffersHandle = static_cast<U32>(-1);
+			U32                      mSystemBuffersHandle = static_cast<U32>(-1);
+
+			TSubMeshRenderInfo       mSubMeshInfo;
+
+			bool                     mIsDirty = true;
+
+#if TDE2_EDITORS_ENABLED
+			std::vector<std::string> mSubmeshesIdentifiers;
+#endif
 	};