+++ TDEngine2/include/graphics/IStaticMeshContainer.h
@@ -17,6 +17,9 @@
 
 namespace TDEngine2
 {
+	struct TSubMeshRenderInfo;
+
+
 	/*!
 		interface IStaticMeshContainer
 
@@ -48,6 +51,8 @@ namespace TDEngine2
 
 			TDE2_API virtual void SetSubMeshId(const std::string& meshName) = 0;
 
+			TDE2_API virtual void SetSubMeshRenderInfo(const TSubMeshRenderInfo& info) = 0;
+
 			/*!
 				\brief The method sets up an internal handle which points to pair 
 				vertex and index buffers within system that renders static meshes
@@ -57,6 +62,12 @@ namespace TDEngine2
 			*/
 
 			TDE2_API virtual void SetSystemBuffersHandle(U32 handle) = 0;
+
+			TDE2_API virtual void SetDirty(bool value) = 0;
+
+#if TDE2_EDITORS_ENABLED
+			TDE2_API virtual void AddSubmeshIdentifier(const std::string& submeshId) = 0;
+#endif
 
 			/*!
 				\brief The method returns an identifier of used material
@@ -82,6 +93,14 @@ namespace TDEngine2
 			*/
 
 			TDE2_API virtual U32 GetSystemBuffersHandle() const = 0;
+
+			TDE2_API virtual const TSubMeshRenderInfo& GetSubMeshInfo() const = 0;
+
+			TDE2_API virtual bool IsDirty() const = 0;
+
+#if TDE2_EDITORS_ENABLED
+			TDE2_API virtual const std::vector<std::string>& GetSubmeshesIdentifiers() const = 0;
+#endif
 		protected:
 			DECLARE_INTERFACE_PROTECTED_MEMBERS(IStaticMeshContainer)
 	};